Remove broken worktree submodule, ignore .claude/worktrees
