+++ package.json
@@ -27,17 +27,14 @@
   "dependencies": {
     "@floating-ui/react": "^0.27.19",
     "@meronex/icons": "^4.0.0",
-    "@modelcontextprotocol/sdk": "^1.29.0",
     "@supabase/supabase-js": "^2.101.1",
     "classnames": "^2.2.6",
-    "express": "^5.2.1",
     "framer-motion": "^12.38.0",
     "lodash-es": "^4.17.20",
     "react": "^19.2.4",
     "react-dom": "^19.2.4",
     "react-textarea-autosize": "^8.3.1",
-    "tailwindcss": "^4.2.2",
-    "zod": "^4.3.6"
+    "tailwindcss": "^4.2.2"
   },
   "devDependencies": {
     "@axe-core/playwright": "^4.11.1",
@@ -45,6 +42,7 @@
     "@tailwindcss/vite": "^4.2.2",
     "@testing-library/dom": "^10.4.1",
     "@testing-library/react": "^16.3.2",
+    "@modelcontextprotocol/sdk": "^1.29.0",
     "@types/express": "^5.0.6",
     "@types/lodash-es": "^4.17.4",
     "@types/react": "^19.2.14",
@@ -59,8 +57,10 @@
     "eslint-plugin-prettier": "^5.0.1",
     "eslint-plugin-react-hooks": "^7.0.1",
     "happy-dom": "^20.8.9",
+    "express": "^5.2.1",
     "prettier": "^3.0.3",
     "tsx": "^4.21.0",
+    "zod": "^4.3.6",
     "typescript": "^6.0.2",
     "vite": "^8.0.3",
     "vite-plugin-pwa": "^1.2.0",
