+++ src/index.tsx
@@ -9,12 +9,12 @@ import Todo from "./components/Todo"
 import Tooltip from "./components/Tooltip"
 import { createRoot } from "react-dom/client"
 import { useStorage } from "./context/StorageContext"
-import { useMCPRelay } from "./hooks/useMCPRelay"
+import { initMCPRelay } from "./hooks/useMCPRelay"
+
+initMCPRelay()
 
 const App = () => {
   const { fetchData } = useStorage()
-
-  useMCPRelay()
 
   React.useEffect(() => {
     fetchData()
